Update mongo connector run scripts to avoid crashing if symlink already exists

+++ mongo-connector_2.5/run.sh
@@ -21,8 +21,13 @@ if [ ! -f "$MONGO_CONNECTOR_CONFIG_FILE_PATH" ] ; then
   echo "$MONGO_CONNECTOR_CONFIG" | sed -e 's/\\\\040/ /g' | python3 -m json.tool > "$MONGO_CONNECTOR_CONFIG_FILE_PATH"
 fi
 
-# Read oplogFile property from config file and create a symlink to /symlink-oplog.timestamp
-ln -s "$(jq --raw-output .oplogFile "$MONGO_CONNECTOR_CONFIG_FILE_PATH")" "$MONGO_CONNECTOR_OPLOG_FILE_SYMLINK_PATH"
+if [ ! -f "$MONGO_CONNECTOR_OPLOG_FILE_SYMLINK_PATH" ] ; then
+  echo "Creating Symlink file: $MONGO_CONNECTOR_OPLOG_FILE_SYMLINK_PATH";
+  # Read oplogFile property from config file and create a symlink to /symlink-oplog.timestamp
+  ln -s "$(jq --raw-output .oplogFile "$MONGO_CONNECTOR_CONFIG_FILE_PATH")" "$MONGO_CONNECTOR_OPLOG_FILE_SYMLINK_PATH"
+else
+  echo "Symlink file: $MONGO_CONNECTOR_OPLOG_FILE_SYMLINK_PATH already exists!";
+fi
 
 echo "Start the app";
 mongo-connector -c "$MONGO_CONNECTOR_CONFIG_FILE_PATH"
